Drishti-Agent-main/streamlit_ui/requirements.txt
streamlit==1.32.0
requests==2.31.0
pandas==2.2.0
plotly==5.18.0
folium>=0.13,<0.15
streamlit-folium==0.16.0
streamlit-option-menu==0.3.6
streamlit-autorefresh==0.0.1
pillow==10.2.0
python-dotenv==1.0.1
